added simulator class
automatically switches what type of simulation is being performed
