+++ GA.py
@@ -1,4 +1,11 @@
-def buildToolbox( individualLength, bitFlipProb, samples):
+
+from deap import base
+from deap import creator
+from deap import gp
+from deap import tools
+from deap import algorithms as algo
+
+def buildToolbox( individualLength, bitFlipProb):
 	# # #setup toolbox
 	toolbox = base.Toolbox()
 	pset = gp.PrimitiveSet("MAIN", arity=1)
@@ -29,118 +36,135 @@ def buildToolbox( individualLength, bitFlipProb, samples):
 	
 	return toolbox, stats
 			
-def setupGAparams(graph, sss):
-	repeat=True
-	while(repeat):
-		repeat=False
-		for node in graph.nodes():
-			if node in graph.successors(node):
-				graph.remove_edge(node,node)
-				repeat=True
-	
-	evaluateNodes=[] #list of nodes that need to be compared to the steady state values (sss)
-	individualParse=[] #list of triples that contain the start and end of the noder order value and operator flag list for each node
-	inputOrderList=[] #list of lists of possible node orders for each node
-	inputOrderInvertList=[] # keeps track of which incoming nodes for each node need to be inverted
-	possibilityNumList=[] # keeps track of length of above inputOrderList for each node
-	earlyEvalNodes=[] # nodes that don't have initial value and need to be re-evaluated early on in the simulation
-	
-	nodeList=graph.nodes()#define the node list simply as the nodes in the graph. 
-	nodeDict={} #identifies names of nodes with their index in the node list
-	for i in range(0,len(nodeList)):
-		nodeDict[nodeList[i]]=i #constructs the node dict so we can easily look up nodes
-	
-	counter=int(0) #keeps track of where we are in the generic individual
-	initValueList=[] #starting states for nodes
-	for j in range(0,len(sss)): #construct empty lists to stick values in later for intiial value list
-			initValueList.append([])
-	
-
-	for i in range(0,len(nodeList)):
-		preds=graph.predecessors(nodeList[i]) # get predecessors of node. 
-		if len(preds)>15: #handle case where there are too many predecessors by truncation
-			preds=preds[1:15]
-		for j in range(0,len(preds)):
-			preds[j]=nodeDict[preds[j]]
-		
-		# the followign section constructs a list of possible node orders
-		withNones = zip(preds, itertool.repeat('empty'))
-		possibilities=list(itertool.product(*withNones))
-		for j in range(0,len(possibilities)):
-			possibilities[j]=list(possibilities[j])
-			while 'empty' in possibilities[j]:
-				possibilities[j].remove('empty')
-			while [] in possibilities[j]:
-				possibilities[j].remove([])
-		while [] in possibilities:
-			possibilities.remove([])
-		
-		#store activities of nodes in correspondence with node order. 
-		activities=[] #list to store activities of nodes (a vs i)
-		for sequence in possibilities:
-			activity=[]
-			for node in sequence:
-				if graph.edge[nodeList[node]][nodeList[i]]['signal']=='a':
-					activity.append(False)
-				else:
-					activity.append(True)
-			activities.append(activity)
-		inputOrderList.append(possibilities)
-		inputOrderInvertList.append(activities)
-		possibilityNumList.append(len(possibilities))
-		
-		#deal with nodes with no incoming or only one incoming first
-		if len(possibilities)==0:
-			individualParse.append([counter,counter,counter])
-		elif len(possibilities)==1:
-			individualParse.append([counter,counter,counter])
-			counter=int(counter+1)
-		else:
-			logNum=ceil(log(len(possibilities))/log(2))#determine number of bits necessary to store information of which node order is the one for that individual
-			individualParse.append([int(counter),int(counter+logNum-1),int(counter+logNum+len(preds)-2)]) #set up limits: lognum-1 for list then len(preds)-1 for the activity flags
-			counter=int(counter+logNum+len(preds)-1) #uptick the counter so we know where to start on the next node
-		#set up lists of initial values and values that need to be evaluated early
-		for j in range(0,len(sss)):
-			ss=sss[j]
-			if  nodeList[i] in sss[0].keys():
-				initValueList[j].append(ss[nodeList[i]])
-				evaluateNodes.append(i)
-			else:
-				initValueList[j].append(0.5)
-				earlyEvalNodes.append(i)
-	return counter-1, individualParse, nodeList, inputOrderList, initValueList, evaluateNodes, possibilityNumList, inputOrderInvertList, earlyEvalNodes
-	
-
-def evaluate(individual, async, iters, sss, evaluateNodes, individualParse, nodeList, inputOrderList, inputOrderInvertList, possibilityNumList, initValueList, steps,earlyEvalNodes,complexPenalty):
+def evaluate(individual, params, model, simulator, sss):
 	SSEs=[]
 	for j in range(0,len(sss)):
 		ss=sss[j]
-		initValues=initValueList[j]
+		initValues=model.initValueList[j]
 		SSE=0
 		if async:
-			boolValues=iterateFuzzyModel(async, iters, individual, individualParse, nodeList, inputOrderList, inputOrderInvertList, possibilityNumList, initValueList[j], steps, 0,0,earlyEvalNodes)	
+			boolValues=simulator.iterateModel(individual, params, model, initValueList[j])	
 		else:
-			boolValues=runFuzzyModel(individual, async, individualParse, nodeList, inputOrderList, inputOrderInvertList, possibilityNumList,initValueList[j], steps,0,0,earlyEvalNodes)	
-		for i in range(0, len(evaluateNodes)):
-			SSE=SSE+(boolValues[evaluateNodes[i]]-ss[nodeList[evaluateNodes[i]]])**2
+			boolValues=simulator.runModel(individual, params, model,initValueList[j])	
+		for i in range(0, len(model.evaluateNodes)):
+			SSE=SSE+(boolValues[model.evaluateNodes[i]]-ss[model.nodeList[model.evaluateNodes[i]]])**2
 		SSEs.append(SSE)
 	edgeDegree=0
 	if complexPenalty:
-		for i in range(0,len(nodeList)):
-			if possibilityNumList[i]>0:
-				edgeDegree=edgeDegree+len(inputOrderList[i][bit2int(individual[individualParse[i][0]:individualParse[i][1]])%possibilityNumList[i]])
+		for i in range(0,len(model.nodeList)):
+			if model.possibilityNumList[i]>0:
+				edgeDegree=edgeDegree+len(model.inputOrderList[i][bit2int(individual[model.individualParse[i][0]:model.individualParse[i][1]])%model.possibilityNumList[i]])
 	SSEs.append(edgeDegree/1.)
-	gc.collect()
 	summer=0.
 	for i in range(0,len(SSEs)):
 		summer+=SSEs[i]
 	return summer,
-
 	
-def writeFuzzyModel(individual, individualParse, nodeList,inputOrderList, inputOrderInvertList, possibilityNumList):
-	#iterate over nodes to generate a BooleanNet representation for the entire model
-	addString=''
-	for i in range(0,len(nodeList)):
-		addString=addString+writeFuzzyNode(i,individual, individualParse, nodeList,inputOrderList, inputOrderInvertList, possibilityNumList)
-		addString=addString+'\n'
-	return addString
+
+def autoSimulator(async, iters, toolbox, stats, graph, trials, sss,individualLength,evaluateNodes, individualParse,nodeList,inputOrderList,inputOrderInvertList,possibilityNumList,initValueList, steps ,nodeNoise,networkNoise,popSize,crossoverProb,mutationProb,generations,earlyEvalNodes, mu,lambd, complexPenalty, genSteps):
+	hofSize=10
+	avgSSEs=[]
+	hofs=[]
+	truthIndividuals=[]
+	truthAvgs=[]
+	hofScores=[]
+	newSSS=[]
+	truthCounter=[0 for number in xrange(hofSize)]
+	trueNodeList=[]
+	for i in range(0,trials):
+		individual=genRandBits(individualLength)
+		newSSS=[]
+		for k in range(0,len(sss)):
+			if async:
+				boolValues=iterateFuzzyModel(async, iters,individual, individualParse, nodeList, inputOrderList, inputOrderInvertList, possibilityNumList, initValueList[k], genSteps,nodeNoise,networkNoise,earlyEvalNodes)
+			else:
+				boolValues=runFuzzyModel(individual, async,individualParse, nodeList, inputOrderList, inputOrderInvertList, possibilityNumList, initValueList[k], genSteps ,nodeNoise,networkNoise,earlyEvalNodes)
+			newSS=copy.deepcopy(sss[k])
+			for j in range(0,len(evaluateNodes)):
+				newSS[nodeList[evaluateNodes[j]]]=boolValues[evaluateNodes[j]]
+			newSSS.append(newSS)
+		#update eval nodes
+		print(evaluate(individual, async, iters, newSSS, evaluateNodes, individualParse, nodeList, inputOrderList, inputOrderInvertList, possibilityNumList, initValueList, steps,earlyEvalNodes,complexPenalty))
+		newInitValueList=[]
+		for j in range(0,len(sss)):
+			newInitValueList.append([])
+		for j in range(0,len(nodeList)):
+			for k in range(0,len(sss)):
+				ss=newSSS[k]
+				if  nodeList[j] in sss[0].keys():
+					newInitValueList[k].append(ss[nodeList[j]])
+				else:
+					newInitValueList[k].append(0.5)
+		#print(len(sss))
+		toolbox.register("evaluate", evaluate, iters=iters, async=async, sss=newSSS, evaluateNodes=evaluateNodes,individualParse= individualParse, nodeList=nodeList, inputOrderList=inputOrderList, inputOrderInvertList=inputOrderInvertList, possibilityNumList=possibilityNumList, initValueList=newInitValueList, steps=steps, earlyEvalNodes=earlyEvalNodes, complexPenalty=complexPenalty)
+		#toolbox.register("evaluate", evaluate, ss=ss, evaluateNodes=evaluateNodes,individualParse= individualParse, nodeList=nodeList, inputOrderList=inputOrderList, inputOrderInvertList=inputOrderInvertList, possibilityNumList=possibilityNumList, initValues=boolValues, steps=steps, h=h,p=p,hillOn=hillOn)
+		population=toolbox.population(n=popSize)
+		hof = tools.HallOfFame(hofSize, similar=numpy.array_equal)
+		output = tools.Logbook()
+		output=algo.eaMuCommaLambda(population, toolbox, mu=mu, lambda_=lambd, stats=stats, cxpb=crossoverProb, mutpb=mutationProb, ngen=generations, verbose=True, halloffame=hof)
+		truth=(writeFuzzyModel(individual, individualParse, nodeList,inputOrderList, inputOrderInvertList, possibilityNumList))
+		
+		for j in range(0,10):
+			bestRun=(writeFuzzyModel(hof[j], individualParse, nodeList,inputOrderList, inputOrderInvertList, possibilityNumList))
+			if truth==bestRun:
+				truthCounter[j]+=1
+				break
+			
+			elif j==0:
+				truthlines=truth.split('\n')
+				newlines=bestRun.split('\n')
+				trueNodes=0
+				for k in range(0,len(truthlines)):
+					if truthlines[k]==newlines[k]:
+						trueNodes=trueNodes+1
+					else:
+						print("incorrect pair: true then test")
+						print(truthlines[k])
+						print(newlines[k])
+				trueNodeList.append(trueNodes)
+		# else:
+			# print('WRONG')
+			# print(truth)
+			# print(bestRun)
+		avgs=[output[1][k]['min'] for k in range(0,len(output[1]))]
+		#print(avgs)
+		hofs.append(hof)
+		temp=[]
+		for hofind in range(0,len(hof)):
+			tempVal=0
+			for bit in range(0,len(hof[hofind])):
+				if not hof[hofind][bit]==individual[bit]:
+					tempVal=tempVal+1
+			temp.append(tempVal)
+		hofScores.append(temp)
+		truthIndividuals.append(individual)
+		truthAvgs.append(boolValues)
+		gc.collect()
+	# prefix='fatimaTest_noise_'+str(nodeNoise)+'_networkNoise_'+str(networkNoise)
+	# outfile = open(prefix+'_hof.pkl', 'wb')
+	# pickle.dump(hof, outfile)
+	# outfile.close()
+	f = open('hof_differences_'+str(nodeNoise)+str(networkNoise)+'.txt', 'w')
+	g = open('hof_individuals'+str(nodeNoise)+str(networkNoise)+'.txt',  'w')
+	h = open('truth_individuals'+str(nodeNoise)+str(networkNoise)+'.txt',  'w')
+	f.write(str(hofScores))
+	f.write('\n')
+	for hofy in range(0,len(hofs)):
+		g.write(str(hofs[hofy]))
+		g.write('\n')
+	h.write(str(truthIndividuals))
+	h.write('\n')
+	f.close()
+	g.close()
+	h.close()
+	print('# true of # trials')
+	for k in range(0,len(truthCounter)):
+		print(truthCounter[k])
+	print(trials)
+	print("for the incorrects, by node")
+	print(trueNodeList)
+	print(len(truthlines))
+	
+if __name__ == '__main__':
+	
+	graph = old.LiuNetwork3Builder()
